+++ src/lib/terrain/heightmap/cpu/geometry-processor.ts
