+++ src/lib/terrain/heightmap/cpu/heightmap-node-geometry.ts
