+++ src/lib/terrain/heightmap/cpu/heightmap-node-id.ts
