+++ src/lib/terrain/heightmap/cpu/heightmap-node-mesh.ts
