+++ src/lib/terrain/heightmap/cpu/heightmap-node.ts
