+++ src/lib/terrain/heightmap/cpu/heightmap-root.ts
