+++ src/lib/terrain/heightmap/cpu/heightmap-viewer-cpu.ts
